Whitespace & move delete icon
