+++ lotti/lib/widgets/journal/entry_detail_linked.dart
@@ -42,70 +42,67 @@ class _LinkedEntriesWidgetState extends State<LinkedEntriesWidget> {
           } else {
             List<String> itemIds = itemsSnapshot.data!;
 
-            return Container(
-              margin: const EdgeInsets.all(8.0),
-              child: Column(
-                children: [
-                  Text(
-                    localizations.journalLinkedEntriesLabel,
-                    style: TextStyle(
-                      color: AppColors.entryTextColor,
-                      fontFamily: 'Oswald',
-                    ),
+            return Column(
+              children: [
+                Text(
+                  localizations.journalLinkedEntriesLabel,
+                  style: TextStyle(
+                    color: AppColors.entryTextColor,
+                    fontFamily: 'Oswald',
                   ),
-                  ...List.generate(
-                    itemIds.length,
-                    (int index) {
-                      String itemId = itemIds.elementAt(index);
+                ),
+                ...List.generate(
+                  itemIds.length,
+                  (int index) {
+                    String itemId = itemIds.elementAt(index);
 
-                      void onDismissed(DismissDirection _) {
-                        String fromId = widget.item.meta.id;
-                        String toId = itemId;
-                        _db.removeLink(fromId: fromId, toId: toId);
-                      }
+                    void onDismissed(DismissDirection _) {
+                      String fromId = widget.item.meta.id;
+                      String toId = itemId;
+                      _db.removeLink(fromId: fromId, toId: toId);
+                    }
 
-                      return Dismissible(
-                        key: ValueKey('Dismissible-$itemId'),
-                        onDismissed: onDismissed,
-                        background: Container(
-                          color: AppColors.error,
-                          child: Row(
-                            mainAxisAlignment: MainAxisAlignment.end,
-                            crossAxisAlignment: CrossAxisAlignment.center,
-                            children: [
-                              Padding(
-                                padding: const EdgeInsets.only(bottom: 6.0),
-                                child: Text(
-                                  localizations.journalUnlinkText,
-                                  style: TextStyle(
-                                    color: AppColors.bodyBgColor,
-                                    fontFamily: 'Oswald',
-                                    fontWeight: FontWeight.w300,
-                                    fontSize: 24,
-                                  ),
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.all(8.0),
-                                child: Icon(
-                                  Icons.link_off,
-                                  size: 32,
+                    return Dismissible(
+                      key: ValueKey('Dismissible-$itemId'),
+                      onDismissed: onDismissed,
+                      background: Container(
+                        color: AppColors.error,
+                        child: Row(
+                          mainAxisAlignment: MainAxisAlignment.end,
+                          crossAxisAlignment: CrossAxisAlignment.center,
+                          children: [
+                            Padding(
+                              padding: const EdgeInsets.only(bottom: 6.0),
+                              child: Text(
+                                localizations.journalUnlinkText,
+                                style: TextStyle(
                                   color: AppColors.bodyBgColor,
+                                  fontFamily: 'Oswald',
+                                  fontWeight: FontWeight.w300,
+                                  fontSize: 24,
                                 ),
                               ),
-                            ],
-                          ),
+                            ),
+                            Padding(
+                              padding: const EdgeInsets.all(8.0),
+                              child: Icon(
+                                Icons.link_off,
+                                size: 32,
+                                color: AppColors.bodyBgColor,
+                              ),
+                            ),
+                          ],
                         ),
-                        child: EntryDetailWidget(
-                          entryId: itemId,
-                          popOnDelete: false,
-                        ),
-                      );
-                    },
-                    growable: true,
-                  )
-                ],
-              ),
+                      ),
+                      child: EntryDetailWidget(
+                        entryId: itemId,
+                        popOnDelete: false,
+                      ),
+                    );
+                  },
+                  growable: true,
+                )
+              ],
             );
           }
         });
