+++ lotti/lib/widgets/journal/entry_details/entry_detail_footer.dart
@@ -6,6 +6,7 @@ import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/theme.dart';
 import 'package:lotti/widgets/journal/duration_widget.dart';
+import 'package:lotti/widgets/journal/entry_details/delete_icon_widget.dart';
 import 'package:lotti/widgets/journal/entry_tools.dart';
 import 'package:lotti/widgets/misc/map_widget.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
@@ -100,6 +101,10 @@ class _EntryDetailFooterState extends State<EntryDetailFooter> {
                           ),
                         ],
                       ),
+                    ),
+                    DeleteIconWidget(
+                      entityId: widget.item.meta.id,
+                      popOnDelete: widget.popOnDelete,
                     ),
                   ],
                 ),
