+++ lotti/lib/widgets/journal/entry_details/entry_detail_header.dart
