+++ lotti/lib/widgets/journal/entry_details/entry_info_row.dart
@@ -6,7 +6,6 @@ import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/logic/persistence_logic.dart';
 import 'package:lotti/theme.dart';
-import 'package:lotti/widgets/journal/entry_details/delete_icon_widget.dart';
 import 'package:lotti/widgets/journal/entry_details/switch_row_widget.dart';
 
 class EntryInfoRow extends StatelessWidget {
@@ -76,10 +75,6 @@ class EntryInfoRow extends StatelessWidget {
                     value: liveEntity.meta.flag == EntryFlag.import,
                   ),
                 ],
-              ),
-              DeleteIconWidget(
-                entityId: entityId,
-                popOnDelete: popOnDelete,
               ),
             ],
           );
